+++ eq/channel.cpp
@@ -77,7 +77,7 @@
 #include "detail/channel.ipp"
 
 #ifdef EQUALIZER_USE_DEFLECT
-#  include "dc/proxy.h"
+#  include "deflect/proxy.h"
 #endif
 
 namespace eq
@@ -234,8 +234,8 @@ const GLEWContext* Channel::glewGetContext() const
 bool Channel::configExit()
 {
 #ifdef EQUALIZER_USE_DEFLECT
-    delete _impl->_dcProxy;
-    _impl->_dcProxy = 0;
+    delete _impl->_deflectProxy;
+    _impl->_deflectProxy = 0;
 #endif
     _impl->framebufferImage.flush();
     return true;
@@ -247,8 +247,8 @@ bool Channel::configInit( const uint128_t& )
     if( getView() &&
         !getView()->getSAttribute( View::SATTR_DISPLAYCLUSTER ).empty( ))
     {
-        LBASSERT( !_impl->_dcProxy );
-        _impl->_dcProxy = new dc::Proxy( this );
+        LBASSERT( !_impl->_deflectProxy );
+        _impl->_deflectProxy = new deflect::Proxy( this );
     }
 #endif
     return true;
@@ -398,9 +398,32 @@ void Channel::frameDrawFinish( const uint128_t&, const uint32_t frameNumber )
 
 void Channel::frameViewStart( const uint128_t& ) { /* nop */ }
 
-void Channel::frameViewFinish( const uint128_t& )
+void Channel::frameViewFinish( const uint128_t& frameID )
 {
+    frameDrawOverlay( frameID );
     _impl->frameViewFinish( *this );
+}
+
+void Channel::frameDrawOverlay( const uint128_t& )
+{
+    applyOverlayState();
+
+#ifdef EQUALIZER_USE_DEFLECT
+    if( _impl->_deflectProxy && _impl->_deflectProxy->isRunning( ))
+    {
+        const eq::PixelViewport& pvp = getPixelViewport();
+        const eq::Viewport& vp = getViewport();
+
+        const float width = pvp.w / vp.w;
+        const float height = pvp.h / vp.h;
+        const float xOffset = vp.x * width;
+
+        glRasterPos3f( 10.f - xOffset, height - 30.f, 0.99f );
+        getWindow()->getMediumFont()->draw( _impl->_deflectProxy->getHelp( ));
+    }
+#endif
+
+    resetOverlayState();
 }
 
 void Channel::setupAssemblyState()
@@ -638,6 +661,33 @@ void Channel::applyOrthoTransform() const
     LB_TS_THREAD( _pipeThread );
     const Matrix4f& xfm = getOrthoTransform();
     EQ_GL_CALL( glMultMatrixf( xfm.array ));
+}
+
+void Channel::applyOverlayState()
+{
+    applyBuffer();
+    applyViewport();
+    setupAssemblyState();
+
+    glMatrixMode( GL_PROJECTION );
+    glLoadIdentity();
+    applyScreenFrustum();
+
+    EQ_GL_CALL( glLogicOp( GL_XOR ));
+    EQ_GL_CALL( glEnable( GL_COLOR_LOGIC_OP ));
+    EQ_GL_CALL( glDisable( GL_DEPTH_TEST ));
+    EQ_GL_CALL( glDisable( GL_LIGHTING ));
+    EQ_GL_CALL( glCullFace( GL_BACK ));
+
+    EQ_GL_CALL( glColor3f( 1.f, 1.f, 1.f ));
+}
+
+void Channel::resetOverlayState()
+{
+    EQ_GL_CALL( glDisable( GL_COLOR_LOGIC_OP ));
+    EQ_GL_CALL( glEnable( GL_DEPTH_TEST ));
+    EQ_GL_CALL( glEnable( GL_LIGHTING ));
+    resetAssemblyState();
 }
 
 namespace
@@ -878,21 +928,11 @@ void Channel::drawStatistics()
         return;
 
     //----- setup
-    EQ_GL_CALL( applyBuffer( ));
-    EQ_GL_CALL( applyViewport( ));
-    EQ_GL_CALL( setupAssemblyState( ));
+    applyOverlayState();
 
-    EQ_GL_CALL( glMatrixMode( GL_PROJECTION ));
-    EQ_GL_CALL( glLoadIdentity( ));
-    applyScreenFrustum();
-
-    EQ_GL_CALL( glMatrixMode( GL_MODELVIEW ));
-    EQ_GL_CALL( glDisable( GL_LIGHTING ));
-
+    EQ_GL_CALL( glDisable( GL_COLOR_LOGIC_OP ));
     EQ_GL_CALL( glEnable( GL_BLEND ));
     EQ_GL_CALL( glBlendFunc( GL_SRC_ALPHA, GL_ONE_MINUS_SRC_ALPHA ));
-    EQ_GL_CALL( glDisable( GL_COLOR_LOGIC_OP ));
-    EQ_GL_CALL( glCullFace( GL_BACK ));
 
 #ifdef EQUALIZER_USE_GLSTATS
     const util::BitmapFont* font = window->getSmallFont();
@@ -907,9 +947,12 @@ void Channel::drawStatistics()
     renderer.draw( data );
 #endif
 
+    EQ_GL_CALL( glEnable( GL_COLOR_LOGIC_OP ));
     EQ_GL_CALL( glColor3f( 1.f, 1.f, 1.f ));
     window->drawFPS();
-    EQ_GL_CALL( resetAssemblyState( ));
+
+    EQ_GL_CALL( glDisable( GL_BLEND ));
+    resetOverlayState();
 }
 
 void Channel::outlineViewport()
@@ -919,8 +962,7 @@ void Channel::outlineViewport()
     if( coreProfile )
         return;
 
-    setupAssemblyState();
-    glDisable( GL_LIGHTING );
+    applyOverlayState();
 
     const eq::PixelViewport& region = getRegion();
     glColor3f( .5f, .5f, .5f );
@@ -940,7 +982,7 @@ void Channel::outlineViewport()
         glVertex3f( pvp.x + .5f,         pvp.getYEnd() - .5f, 0.f );
     } glEnd();
 
-    resetAssemblyState();
+    resetOverlayState();
 }
 
 namespace detail