+++ eq/deflect/proxy.h
@@ -15,8 +15,8 @@
  * 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
  */
 
-#ifndef EQ_DC_PROXY_H
-#define EQ_DC_PROXY_H
+#ifndef EQ_DEFLECT_PROXY_H
+#define EQ_DEFLECT_PROXY_H
 
 #include <eq/resultImageListener.h> // base class
 #include <eq/types.h>
@@ -24,22 +24,20 @@
 
 namespace eq
 {
-namespace dc
+namespace deflect
 {
-namespace detail { class Proxy; }
-
 /** @internal */
 class Proxy : public ResultImageListener
 {
 public:
-    /** Construct a DisplayCluster proxy associated to a destination channel. */
+    /** Construct a Deflect proxy associated to a destination channel. */
     explicit Proxy( Channel* channel );
 
-    /** Destruct the DisplayCluster proxy. */
+    /** Destruct the Deflect proxy. */
     ~Proxy();
 
-    /** Stream the given image to DisplayCluster. */
-    void notifyNewImage( eq::Channel& channel, const eq::Image& image ) final;
+    /** Stream the given image to the Deflect host. */
+    void notifyNewImage( Channel& channel, const Image& image ) final;
 
     /** @return the associated destination channel. */
     Channel* getChannel();
@@ -47,24 +45,40 @@ public:
     /** @return the underlying socket descriptor. */
     int getSocketDescriptor() const;
 
-    /** @return true if a new Event was sent by DisplayCluster. */
+    /** @return true if a new Event was sent by the Deflect host. */
     bool hasNewEvent() const;
 
-    /** @return whether the application is running in DisplayCluster. */
+    /** @return whether the application is running on the Deflect host. */
     bool isRunning() const;
 
     /**
      * Signal the proxy to stop running.
      *
-     * This is called if the application is closed in DisplayCluster.
+     * This is called if the application is closed in the Deflect host.
      */
     void stopRunning();
 
     /** @return the latest window Event. @sa hasNewEvent() */
-    deflect::Event getEvent() const;
+    ::deflect::Event getEvent() const;
+
+    enum NavigationMode
+    {
+        MODE_PAN,
+        MODE_ROTATE
+    };
+
+    /** Set the navigation mode for received mouse events. */
+    void setNavigationMode( NavigationMode mode );
+
+    /** @return the current set navigation mode. */
+    NavigationMode getNavigationMode() const;
+
+    /** @return a help text which can be printed as an overlay. */
+    std::string getHelp() const;
 
 private:
-    detail::Proxy* const _impl;
+    class Impl;
+    std::unique_ptr< Impl > _impl;
 };
 }
 }